deepfakeapp/requirements.txt
streamlit==1.52.2
google-genai==1.56.0
Pillow==11.2.1
PyMuPDF==1.26.7
feedparser==6.0.12
requests==2.32.5
beautifulsoup4==4.14.3